3rd_Checking if a button is released
- "BallPressed.cs" was renamed to "BallHandler.cs". This script checks a button's stauts of being pressed and being released.
- "BallController.cs" has been modified for test reason.
- If a user press the button, "BallHandler.cs" gets the player game object's inner script and call movement-related method in it.
- If the button is released, "BallHandler.cs" does the same thing with the case above, but call stop-related method in it.
